base.py: Fix FAIL creation in execute()
The execute() return fail if the file named FAIL exists and always
remove the FAIL at the end. If there are two dependent execute(A)
and execute(B), B's fail does not make A fail as expected because
the FAIL had been removed by B.

Add datetime to create different fail files for each command
execution.