base.py: Fix fail file name in execute()
The fail file is named based on local time, if execute() is called two
times in one second, they will create the fail file with same name, and
the second execute will delete the fail file of the first one, which
causes the first execute cannot return correct value.

Use uuid instead of time for the fail file name.