base.py: Update webgpu-cts expectations to adapt to all Intel GPUs
Currently we just update the expecations with 'intel-gen-9' to
'intel-gen-12', but more tags like 'intel-<device-id>' are used in the
expectations.txt, and we're not sure how many device ids will be added.

To prevent the code from being updated every time when a new device id
or genertation is added, we update the expectations with specific intel
gpu tag 'intel-*' to 'intel' to match all intel platforms.